pyproject.toml
[build-system]
requires = ["setuptools>=44", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observatory-platform"
version = "1.0.0"
description = "The Observatory Platform is an environment for fetching, processing and analysing data to understand how well universities operate as Open Knowledge Institutions."
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["science", "data", "workflows", "academic institutes", "academic-observatory-workflows"]
authors = [{ name = "Curtin University", email = "agent@observatory.academy" }]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
dependencies = [
    # Airflow
    "apache-airflow[redis,celery,slack,http]==2.7.3",
    # Google Cloud
    "google-crc32c>=1.1.0,<2",
    "google-cloud-bigquery>=3,<4",
    "google-api-python-client>=2,<3",
    "google-cloud-storage>=2.7.0,<3",
    "google-auth-oauthlib>=0.4.5,<1",
    "google-cloud-compute >=1.16.0,<2.0",
    # AWS
    "boto3>=1.15.0,<2",
    # Command line interface, config files and templates
    "click>=8,<9",
    "pyyaml>=6,<7",
    "cerberus>=1.3.4,<2",
    "Jinja2>=3,<4",
    "stringcase>=1.2.0,<2",
    "pyhcl>=0.4.4,<1",
    # Reading and writing jsonlines files
    "jsonlines>=2.0.0,<3", # Writing
    "json_lines>=0.5.0,<1", # Reading, including jsonl.gz
    # HTTP requests and URL cleaning
    "requests>=2.25.0,<3",
    "tldextract>=3.1.1",
    "aiohttp>=3.7.0,<4",
    "responses>=0.23.1,<1",
    # SFTP
    "pysftp>=0.2.9,<1",
    "paramiko>=2.7.2,<3",
    # FTP
    "pyftpdlib>=1.5.7,<2",
    # Utils
    "natsort>=7.1.1,<8",
    "backoff>=2,<3",
    "timeout-decorator",
    "validators<=0.20.0",
    "xmltodict",
    "pandas",
    "tenacity",
    # Test utils
    "time_machine>=2.0.0,<3",
    "httpretty>=1.0.0,<2",
    "sftpserver>=0.3,<1",
    # Backports
    "typing-extensions>=3.10.0.1,<4; python_version<'3.10'",
    "oauth2client>=4.1.0,<5",
    "deprecation>2,<3",
    "virtualenv",
    "deepdiff",
]

[project.optional-dependencies]
test = [
    "liccheck>=0.4.9,<1",
    "flake8>=3.8.0,<4",
    "coverage>=5.2,<6",
    "faker>=8.12.1,<9",
    "redis>=3.5.3,<4",
    "boto3>=1.15.0,<2",
    "azure-storage-blob>=12.8.1,<13"
]

[project.urls]
"Homepage" = "https://github.com/The-Academic-Observatory/observatory-platform"
"Bug Tracker" = "https://github.com/The-Academic-Observatory/observatory-platform/issues"
"Documentation" = "https://observatory-platform.readthedocs.io/en/latest/"
"Source" = "https://github.com/The-Academic-Observatory/observatory-platform"
"Repository" = "https://github.com/The-Academic-Observatory/observatory-platform"
